% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.zerorow}
\alias{.zerorow}
\title{Returns an element suitable for inserting into a column of a zero-row dataframe.}
\usage{
.zerorow(x)
}
\arguments{
\item{x}{The value for which you desire a zero-length version.}
}
\description{
\itemize{
\item If \code{x} is a dataframe itself, the zero-row version of that dataframe is returned.
\item If \code{x} is a vector, then \code{x[0]} is returned.
\item If \code{x} is something else, then \code{x[0]} is returned, possibly unsafely.
}
}
