% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepend_class}
\alias{prepend_class}
\title{Prepend classname(s) to an object's S3 class attribute.}
\usage{
prepend_class(obj, klass)
}
\arguments{
\item{obj}{an single object.}

\item{klass}{a character vector of classnames, ordered appropriately (following S3 OO semantics).}
}
\value{
the object with \code{klass} classnames prepended to the \code{class} attribute of \code{obj}.
}
\description{
Prepend classname(s) to an object's S3 class attribute.
}
