% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbldef.R
\name{tbldef}
\alias{tbldef}
\alias{tbldef.col}
\title{Table definition/specification.}
\usage{
tbldef.col(cast = ident, ptype = NULL)

tbldef(coldefs = list(), klass = NULL)
}
\arguments{
\item{cast}{A casting function that casts its first argument into the appropriate type.
Ideally, this function should handle NULL and convert it into the appropriate type, but if not, use \code{ptype} for more precise control.
See details for more information.}

\item{ptype}{The 'prototype' value to use when a column is missing as part of a cast procedure. See details.}

\item{coldefs}{A named list of \code{tbldef.col} values.}

\item{klass}{A classname (or ordered classnames, if more than one) to prepend during the table-casting process.}
}
\description{
Constructs a table specification.
}
\details{
\subsection{\code{tbldef.col}}{

Each table's definition is made up of a \emph{named} list of \code{tbldef.col} definitions, specifying each column's type.
\code{tbldef.col} takes a casting function \code{cast} and optionally a \code{ptype} (the latter of which is described below in "Missing columns").
The default \code{cast} argument is \code{ident}, which is the identify function and performs no casting on the column at all.
}

\subsection{\verb{[0]}-ability}{

The table's \code{cast} procedure uses row-binding approach and pre-computes a zero-row table (the definition's \code{ptype}) to which a (possibly modified) input table will row-combined.
To compute the zero-row table, the results of each column's \code{cast} function must be indexable such that it can be truncated to a length-zero vector or list (or a zero-row dataframe, in the case of a folded dataframe column).
Generally-speaking, that means each column's \code{cast} and \code{ptype} argument's should be callable like so: \code{cast(ptype)[0]} or (\code{cast(ptype)[0,]} in the folded dataframe case).
}

\subsection{Missing columns}{

When an input table is missing columns that are part of the table definition, those missing columns are added to the output.
The column type in that case is the result of calling the column's \code{cast} function on the \code{ptype} argument.
By default \code{ptype} is \code{NULL}, so \code{cast} should be tolerant of \code{NULL} and return a non-\code{NULL} result, as \code{NULL} values cannot be in a dataframe column (apart from elements in a listcol).
If the \code{cast} function is not tolerant of \code{NULL}, specify the \code{ptype} such that the column's \code{cast} returns an appropriately-typed value.
}

\subsection{'Folded' dataframes}{

Tibbles support the notion of a column-collapsed 'folded' dataframe.
This is distinct from, but conceptually related-to, 'nested' dataframes and listcols.
In the folded dataframe case, there is no listcol: the type of the column itself is \code{tbl}.
Folded dataframes have the same number of rows as the surrounding dataframe of which they are a part, but simply allow the columns to be nested.
For example:\if{html}{\out{<div class="r">}}\preformatted{tibble::tibble(col1 = 1:3, col2 = tibble::tibble(col1 = 3:1, col2 = 1:3))
}\if{html}{\out{</div>}}

In that example, the 'outer' table has two columns (\code{col1} and \code{col2}), while the inner table (\code{col2}) has two columns (also called \code{col1} and \code{col2}, but distinct from the 'parent' columns of the same name).
There are only three 'vector' columns in the full table, however: \code{col1}, \code{col2$col1}, and \code{col2$col2}.

\code{tbldef} supports this notion by permitting additional \code{tbldef}s as values within the \code{coldefs} argument.
}

\subsection{Return values}{

\code{tbldef.col} returns a structure that should only be used within the context of a \code{tbldef} (hence the naming scheme).
\code{tbldef} returns a list with three components:
\itemize{
\item \code{cast} the casting function with this signature: \code{cast(x = data.frame(), .keep_all = TRUE, ...)}.
\itemize{
\item \code{x} is the table on which to perform the casting. \code{NULL} is treated identically to an empty (0 x 0) table.
\item \code{.keep_all} is a logical flag telling the casting function whether or not \emph{additional} columns in \code{x} should be retained.
These are columns that are not part of the table definition, but perhaps one wants to keep attached to the output table.
In such cases, these columns are moved to the end (right side) and are otherwise not touched.
}
\item \code{ptype} a zero-row 'prototype' table with typed columns.
\item \code{klass} simply a pass-back of the \code{klass} argument in the \code{tbldef} function (for reference when managing table definitions).
}
}
}
\examples{
def <- tbldef(list(
  col1 = tbldef.col(as.integer),
  col2 = tbldef.col(as.character)
))

## casting an empty table (or NULL) results in the zero-row ptype:
def$cast()
def$ptype

## cast a table with a missing column and drop an extraneous column:
def$cast(data.frame(col1 = c("1", "2", "3"), col3 = c(TRUE, FALSE, TRUE)))

## now let's do the same but keep the extraneous column and demonstrate its repositioning:
def$cast(data.frame(col3 = c(TRUE, FALSE, TRUE), col1 = c("1", "2", "3")), .keep_all = TRUE)

}
